rag-service/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pydantic
python-dotenv
loguru
openai
tenacity
